Updating to work with Amazon linux 2023

+++ bootstrap_instance.sh
@@ -10,7 +10,13 @@ cat <<EOF >> ~/.bashrc
 source "$SCRIPT_DIR/shell/common_profile.sh"
 EOF
 
-sudo amazon-linux-extras install epel -y
+release_version=$(cat /etc/system-release)
+AL2023_VERSION="Amazon Linux release 2023 (Amazon Linux)"
+
+if [[ $release_version != $AL2023_VERSION ]]; then
+  # 2023 doesn't support epel
+  sudo amazon-linux-extras install epel -y
+fi
 
 sudo yum install -y \
   git \
@@ -26,8 +32,12 @@ sudo yum install -y \
   openssl-devel \
   tk-devel \
   libffi-devel \
-  xz-devel \
-  ShellCheck
+  xz-devel
+
+if [[ $release_version != $AL2023_VERSION ]]; then
+  # 2023 doesn't support epel so I can't figure out how to install shellcheck
+  yum install -y ShellCheck
+fi
 
 curl https://pyenv.run | bash
 
@@ -35,9 +45,11 @@ export PYENV_ROOT="$HOME/.pyenv"
 [[ -d $PYENV_ROOT/bin ]] && export PATH="$PYENV_ROOT/bin:$PATH"
 eval "$(pyenv init -)"
 
-# Fix for amazon linux 2
-sudo yum remove openssl-devel -y
-sudo yum install openssl11-devel -y
+if [[ $release_version != $AL2023_VERSION ]]; then
+  # Fix for amazon linux 2
+  sudo yum remove openssl-devel -y
+  sudo yum install openssl11-devel -y
+fi
 
 pyenv install 3.11
 pyenv global 3.11
